+++ bigquery/src/test/java/com/google/cloud/hadoop/io/bigquery/BigQueryHelperTest.java
@@ -185,6 +185,7 @@ public class BigQueryHelperTest {
         /* timePartitioning= */ null,
         kmsKeyName,
         BigQueryFileFormat.NEWLINE_DELIMITED_JSON,
+        BigQueryConfiguration.OUTPUT_TABLE_CREATE_DISPOSITION_DEFAULT,
         BigQueryConfiguration.OUTPUT_TABLE_WRITE_DISPOSITION_DEFAULT,
         ImmutableList.of("test-import-path"),
         true);