+++ bigquery/CHANGES.md
@@ -36,6 +36,10 @@
 
         mapred.bq.dynamic.file.list.record.reader.poll.max.attempts (default: -1)
 
+1. Add a property to specify output table create disposition:
+
+       mapred.bq.output.table.createdisposition (default: CREATE_IF_NEEDED)
+
 ### 0.13.14 - 2019-02-13
 
 1.  POM updates for GCS connector 1.9.14.